backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv

pytz

pytz
supabase
onnxruntime
livekit

openai
livekit-api

livekit-plugins-openai
livekit-plugins-deepgram 
livekit-plugins-noise-cancellation
livekit-plugins-silero
omegaconf
torch
torchaudio
numpy
scipy